Removed a number of media queries if Cordoba - they cause lots of problems Split the CordovaUtilites into two files -- new CordovaOffsets.jsx Pulled the inline BallotTitleHeader code out of Ballot and Vote, it was the same code, and would have needed three level ternary in each instance. Adjusted the height of the Values page for all Android.
