+++ src/js/components/Intro/IntroNetworkDefinition.jsx
@@ -1,7 +1,8 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { renderLog } from '../../utils/logging';
-import { cordovaDot, cordovaNetworkNextButtonTop } from '../../utils/cordovaUtils';
+import { cordovaNetworkNextButtonTop } from '../../utils/cordovaOffsets';
+import { cordovaDot } from '../../utils/cordovaUtils';
 
 /*
 The problem with urls in css for Apache Cordova
