+++ src/js/components/Navigation/FooterBar.jsx
@@ -8,7 +8,8 @@ import BallotIcon from '@material-ui/icons/Ballot';
 import HowToVoteIcon from '@material-ui/icons/HowToVote';
 // import PeopleIcon from '@material-ui/icons/People'; // DALE: FRIENDS TEMPORARILY DISABLED
 import styled from 'styled-components';
-import { cordovaFooterHeight, historyPush } from '../../utils/cordovaUtils';
+import { cordovaFooterHeight } from '../../utils/cordovaOffsets';
+import { historyPush } from '../../utils/cordovaUtils';
 import { stringContains } from '../../utils/textFormat';
 import FriendStore from '../../stores/FriendStore';
