+++ src/js/components/Navigation/WelcomeAppbar.jsx
@@ -12,7 +12,8 @@ import MenuIcon from '@material-ui/icons/Menu';
 import cookies from '../../utils/cookies';
 import Navigation, { LogoContainer, Divider, NavLink, MobileNavigationMenu, MobileNavDivider, NavRow } from '../Welcome/Navigation';
 import HeaderBarLogo from './HeaderBarLogo';
-import { historyPush, cordovaWelcomeAppToolbarTop } from '../../utils/cordovaUtils';
+import { cordovaWelcomeAppToolbarTop } from '../../utils/cordovaOffsets';
+import { historyPush } from '../../utils/cordovaUtils';
 import AppStore from '../../stores/AppStore';
 import AppActions from '../../actions/AppActions';
 import HeaderBarProfilePopUp from './HeaderBarProfilePopUp';
