+++ src/js/routes/Ballot/Ballot.jsx
@@ -12,6 +12,7 @@ import BallotActions from '../../actions/BallotActions';
 import BallotElectionList from '../../components/Ballot/BallotElectionList';
 import BallotDecisionsTabs from '../../components/Navigation/BallotDecisionsTabs';
 import BallotItemCompressed from '../../components/Ballot/BallotItemCompressed';
+import BallotTitleHeader from './BallotTitleHeader';
 import BallotSideBar from '../../components/Navigation/BallotSideBar';
 import BallotSearch from '../../components/Ballot/BallotSearch';
 import BallotStatusMessage from '../../components/Ballot/BallotStatusMessage';
@@ -19,9 +20,8 @@ import BallotStore from '../../stores/BallotStore';
 import BallotSummaryModal from '../../components/Ballot/BallotSummaryModal';
 import BrowserPushMessage from '../../components/Widgets/BrowserPushMessage';
 import cookies from '../../utils/cookies';
-import {
-  cordovaBallotFilterTopMargin, historyPush, cordovaScrollablePaneTopPadding, isCordova, isWebApp,
-} from '../../utils/cordovaUtils';
+import { cordovaBallotFilterTopMargin, cordovaScrollablePaneTopPadding } from '../../utils/cordovaOffsets';
+import { historyPush, isCordova, isWebApp } from '../../utils/cordovaUtils';
 import ElectionActions from '../../actions/ElectionActions';
 import ElectionStore from '../../stores/ElectionStore';
 import isMobile from '../../utils/isMobile';
@@ -782,6 +782,7 @@ class Ballot extends Component {
     });
   }
 
+
   render () {
     // console.log('Ballot render');
     renderLog(__filename);
@@ -815,11 +816,11 @@ class Ballot extends Component {
     const voterAddressMissing = this.state.location === null;
 
     // const ballot_caveat = BallotStore.ballotProperties.ballot_caveat; // ballotProperties might be undefined
-    const electionName = BallotStore.currentBallotElectionName;
+    const electionName = BallotStore.currentBallotElectionName || '';
     const electionDayText = BallotStore.currentBallotElectionDate;
     const sourcePollingLocationWeVoteId = BallotStore.currentBallotPollingLocationSource;
     const ballotReturnedAdminEditUrl = `${webAppConfig.WE_VOTE_SERVER_ROOT_URL}b/${sourcePollingLocationWeVoteId}/list_edit_by_polling_location/?google_civic_election_id=${VoterStore.electionId()}&state_code=`;
-    // console.log("electionName: ", electionName, ", electionDayTextFormatted: ", electionDayText);
+    // console.log('electionName: ', electionName, ', electionDayText: ', electionDayText);
 
     const emptyBallotButton = completionLevelFilterType !== 'none' && !voterAddressMissing ? (
       <span>
@@ -853,7 +854,8 @@ class Ballot extends Component {
       </div>
     ) : null;
 
-    const electionDayTextFormatted = electionDayText ? <span>{moment(electionDayText).format('MMM Do, YYYY')}</span> : <span />;
+    const electionDayTextFormatted = electionDayText ? moment(electionDayText).format('MMM Do, YYYY') : '';
+    // console.log('electionName: ', electionName, ', electionDayTextFormatted: ', electionDayTextFormatted);
 
     const inRemainingDecisionsMode = completionLevelFilterType === 'filterRemaining';
 
@@ -874,20 +876,7 @@ class Ballot extends Component {
                 <div className="col-md-12">
                   <Helmet title="Ballot - We Vote" />
                   <header className="ballot__header__group">
-                    <h1 className={isCordova() ?  'ballot__header__title__cordova' : 'ballot__header__title'}>
-                      { electionName ? (
-                        <span className={isWebApp() ? 'u-push--sm' : 'ballot__header__title__cordova-text'}>
-                          {electionName}
-                          {' '}
-                          <span className="d-none d-sm-inline">&mdash; </span>
-                          <span className="u-gray-mid u-no-break">{electionDayTextFormatted}</span>
-                        </span>
-                      ) : (
-                        <span className="u-push--sm">
-                         Loading Election...
-                        </span>
-                      )}
-                    </h1>
+                    <BallotTitleHeader electionName={electionName} electionDayTextFormatted={electionDayTextFormatted} />
                   </header>
 
                   { textForMapSearch || ballotWithItemsFromCompletionFilterType.length > 0 ? (
