+++ src/js/routes/More/Pricing.jsx
@@ -5,7 +5,8 @@ import PropTypes from 'prop-types';
 import styled from 'styled-components';
 import { withStyles } from '@material-ui/core/styles';
 import AppActions from '../../actions/AppActions';
-import { cordovaScrollablePaneTopPadding, historyPush } from '../../utils/cordovaUtils';
+import { cordovaScrollablePaneTopPadding } from '../../utils/cordovaOffsets';
+import { historyPush } from '../../utils/cordovaUtils';
 import { renderLog } from '../../utils/logging';
 import Footer from '../../components/Welcome/Footer';
 import Section from '../../components/Welcome/Section';
