+++ src/js/routes/WelcomeForCampaigns.jsx
@@ -7,7 +7,8 @@ import { withStyles } from '@material-ui/core/styles';
 import AnalyticsActions from '../actions/AnalyticsActions';
 import AppActions from '../actions/AppActions';
 import Footer from '../components/Welcome/Footer';
-import { historyPush, cordovaDot, cordovaScrollablePaneTopPadding } from '../utils/cordovaUtils';
+import { cordovaScrollablePaneTopPadding } from '../utils/cordovaOffsets';
+import { historyPush, cordovaDot } from '../utils/cordovaUtils';
 import Testimonial from '../components/Widgets/Testimonial';
 import validateEmail from '../utils/email-functions';
 import VoterActions from '../actions/VoterActions';
