+++ src/js/routes/WelcomeForOrganizations.jsx
@@ -7,7 +7,8 @@ import { withStyles } from '@material-ui/core/styles';
 import AnalyticsActions from '../actions/AnalyticsActions';
 import AppActions from '../actions/AppActions';
 import Footer from '../components/Welcome/Footer';
-import { historyPush, cordovaDot, cordovaScrollablePaneTopPadding } from '../utils/cordovaUtils';
+import { cordovaScrollablePaneTopPadding } from '../utils/cordovaOffsets';
+import { historyPush, cordovaDot } from '../utils/cordovaUtils';
 import { renderLog } from '../utils/logging';
 import Testimonial from '../components/Widgets/Testimonial';
 import validateEmail from '../utils/email-functions';
