+++ src/js/routes/WelcomeForVoters.jsx
@@ -10,7 +10,8 @@ import EmailIcon from '@material-ui/icons/Email';
 import { Title, BlueTitle, SubTitle, Video, PlayerContainer } from '../components/Welcome/HeaderWelcome';
 import AddressBoxWelcome from '../components/Welcome/AddressBoxWelcome';
 import AnalyticsActions from '../actions/AnalyticsActions';
-import { historyPush, cordovaDot, cordovaScrollablePaneTopPadding } from '../utils/cordovaUtils';
+import { cordovaScrollablePaneTopPadding } from '../utils/cordovaOffsets';
+import { historyPush, cordovaDot } from '../utils/cordovaUtils';
 import Footer from '../components/Welcome/Footer';
 import { renderLog } from '../utils/logging';
 import TextBox from '../components/Welcome/TextBox';
